requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
redis
python-jose
celery
joblib
scikit-learn
pytest
pandas